extracted the sensor status code into it's own module
